shipper: models: patch size field to BigIntegerField
The current IntegerField limits the size of builds to 2.1GB. Oops!

+++ shipper/models.py
@@ -158,7 +158,7 @@ class Build(models.Model):
         max_length=500,
         unique=True,
         help_text="Example: 'Bliss-v14-bullhead-OFFICIAL-gapps-20200608")
-    size = models.IntegerField(
+    size = models.BigIntegerField(
         help_text="Size of zip file in bytes<br>Example: 857483855"
     )
     version = models.TextField(
